Parsed out html into templates
The original container all the divs on the page to begin with and just
hid and unhid as needed.  As practice and to avoid loading things not
needed immediately, I moved the individual "pages" into templates that
are loaded as needed.  Once loaded, they can then be hidden as needed.